tokenizer実装OK
エッジケースとかエラーになるケースは後で戻ってきてから修正する
