+++ tokenizer/Makefile
@@ -4,8 +4,10 @@ SRCS_DIR := ./
 OBJS_DIR := ./objs
 
 SRCS := main.c \
+		tokenizer.c \
 		tokenizer_helper.c \
 		tokenizer_utils.c \
+		utils.c \
 
 OBJS := $(SRCS:.c=.o)
 SRCS := $(addprefix $(SRCS_DIR)/, $(SRCS))
