+++ tokenizer/tokenizer.h
@@ -30,7 +30,11 @@ bool	is_blank(char c);
 bool	is_meta_character(char c);
 bool	is_operator_charcter(char c);
 
+t_token	*tokenize(char *input, bool *is_err);
+
 t_token	*new_token(t_token *cur, t_token_type type, char *literal);
 void	skip_blanks(char **begin);
+
+void	perror_exit(char *func_name);
 
 #endif
